requirements.txt
streamlit>=1.30.0
pandas>=2.1.0
openpyxl>=3.1.2
pdfplumber>=0.10.3
pdfkit>=1.0.0
Jinja2>=3.1.3
python-dotenv>=1.0.0
google-genai>=0.3.0
